refactor(web): refineadvocates hook
